Interaction editing & Unit Test
* Interactions work better now, but now we need to make sure that the few edge cases work
