+++ src/plugins/mod.rs
@@ -4,6 +4,7 @@ mod player;
 mod debug;
 mod tilemap;
 mod mob;
+mod interactions;
 
 // Make components & systems visible to program
 pub use pig::*;
@@ -12,3 +13,4 @@ pub use player::*;
 pub use debug::*;
 pub use tilemap::*;
 pub use mob::*;
+pub use interactions::*;
